% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twopiece.R
\name{dtp4}
\alias{dtp4}
\title{Probability Density Function for 4-parameter twopiece distributions}
\usage{
dtp4(x, mu, par1, par2, delta, FUN, param = "tp", log = FALSE)
}
\arguments{
\item{log, }{log.p: logical; if TRUE, probabilities p are given as log(p).}

\item{x:}{vector of quantiles.}

\item{p:}{vector of probabilities.}

\item{n:}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{mu:}{location parameter.}

\item{par1:}{scale parameter 1.}

\item{par2:}{scale parameter 2.}

\item{delta:}{shape parameter.}

\item{FUN:}{a symmetric density f.}

\item{param:}{parameterizations used.}
}
\description{
Probability Density Function for 4-parameter twopiece distributions
}
