% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmixnorm.R
\name{rmixnorm}
\alias{rmixnorm}
\title{Multivariate draws from Normal mixture}
\usage{
rmixnorm(n, mu, Sigma, probs)
}
\arguments{
\item{n}{the number of draws}

\item{mu}{a list of means}

\item{Sigma}{a list of convariance matrices for the distributions}

\item{probs}{a numeric non-negative vector of length K, specifying
the probability for the K classes}
}
\description{
This function generates a specified number of draws from
a mixture of normal distributions
}
