% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twopiece-package.R
\docType{package}
\name{twopiece-package}
\alias{twopiece}
\alias{twopiece-package}
\title{Two Piece}
\description{
To be written
}
\author{
\strong{Maintainer}: Francisco Javier Rubio \email{fxrubio@gmail.com}

Authors:
\itemize{
  \item A. M. Lopez \email{amonteslop@gmail.com}
  \item David Rossell \email{rosselldavid@gmail.com}
  \item Michael DeWitt \email{me.dewitt.jr@gmail.com} (\href{https://orcid.org/0000-0001-8940-1967}{ORCID})
}

}
\keyword{internal}
