% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmixnorm.R
\name{dmixnorm}
\alias{dmixnorm}
\title{Density for mixture of Normals}
\usage{
dmixnorm(x, mu, Sigma, probs, logscale = FALSE)
}
\arguments{
\item{x}{vector or matrix of quantiles. If x is a matrix, each row is taken
to be a quantile.}

\item{mu}{a list of means}

\item{Sigma}{a list covariance matrices}

\item{probs}{a numeric non-negative vector of length K, specifying the
probability for the K classes}

\item{logscale}{return log density if \code{TRUE}, default \code{FALSE}}
}
\description{
This function outputs the density for a mixture of normal distribution
}
