% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtp4.R
\name{dtp4}
\alias{dtp4}
\alias{qtp4}
\alias{rtp4}
\alias{ptp4}
\title{The 4-Parameter Two Piece Distribution}
\usage{
dtp4(x, mu, par1, par2, delta, FUN, param = "tp", log = FALSE)
ptp4(x, mu, par1, par2, delta, FUN, param = "tp", log.p = FALSE)
qtp4(p, mu, par1, par2, delta, FUN, param = "tp")
rtp4(n, mu, par1, par2, delta, FUN, param = "tp")
}
\arguments{
\item{x}{vector of quantiles}

\item{mu}{location parameter, \eqn{\mu}}

\item{par1}{scale parameter 1, \eqn{\sigma_1}}

\item{par2}{scale parameter 2, \eqn{\sigma_2}}

\item{delta}{shape parameter, \eqn{\delta}}

\item{FUN}{a symmetric density \code{f}}

\item{param}{parameterizations used with default of \code{"tp"}}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}

\item{p}{vector of probabilities}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p)}
}
\description{
Density, distribution function, quantile function and random generation for
the 4-parameter two piece distribution with 3 parameterizations:
two-piece (tp), epsilon-skew (eps), and inverse scale factors (isf)
}
\details{
The 4-parameter two piece distribution with parameters \eqn{\mu}
\eqn{\sigma_1}and \eqn{\sigma_2} and \eqn{\delta}
has the following density:

\deqn{s(x) = \frac{2}{\sigma_1 + \sigma_2}f( ( x -\mu )/\sigma_1, \delta ) \quad for \quad x<\mu }
and
\deqn{s(x) = \frac{2}{\sigma_1 + \sigma_2}f( ( x -\mu )/\sigma_2, \delta ) \quad for \quad x\geq \mu }

where \code{f(x)} is a symmetric density about zero.

If param is not specified, it assumes the default value of "tp".
Information about the "eps" and "isf" parameterizations can be found in the References.
\code{dtp4} gives the density
\code{ptp4} gives the distribution function
\code{qtp4} gives the quantile function and \code{rtp4} generates random deviates
}
\examples{

## 4-parameter two piece Student-t density with parameterization 'tp'
tempf = function(x) dtp4(x,0,3,1,4,dt,param="tp")
curve(tempf,-10,5)

## 4-parameter two piece Student-t distribution with parameterization 'tp'
tempf = function(x) ptp4(x,0,3,1,4,pt,param="tp")
curve(tempf,-10,10)

## random number generation for 4-parameter two piece Student-t distribution
## with parameterization 'tp'
sim <- rtp4(1000,0,1,1,10,rt)
hist(sim, probability=TRUE, xlim=c(-10,10),ylim=c(0,dt(0,4)))

## quantile function for the 4-parameter two piece Student-t distribution
## with parameterization 'tp'
qtp4(0.5, 0, 1, 1, 4, qt ,param = "tp")
}
\references{
Arellano-Valle, R. B Gómez, H. W. and Quintana, F. A. (2005). Statistical inference for general class of
asymmetric distributions. \emph{Journal of Statistical Planning and Inference}, \bold{128}: 427-443.

Fernández, C. and Steel, M. F. J. (1998). On Bayesian modeling of fat tails and skewness.
\emph{Journal of the American Statistical Asociation}, \bold{93}, 359-371.

Mudholkar, G. S. and Hutson, A. D. (2000). The epsilon-skew-normal distribution for analyzing
near-normal data. \emph{Journal of Statistical Planning and Inference}, \bold{83}: 291-309.

Rubio, F. J. and Steel, M. F. J. (2014). Inference in Two-Piece Location-Scale models with
Jeffreys Priors, with discussion. \emph{Bayesian Analysis}, \bold{9}: 1-22.
}
\seealso{
\link{dnorm} for the normal distribution and \link{dt} for the Student t distribution
\code{\link{dtp3}} for the 3-parameter two piece distribution.
}
\author{
F. J. Rubio \email{fxrubio@gmail.com} and A. M. López \email{amonteslop@gmail.com}
}
