% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twopiece.R
\name{rtp3}
\alias{rtp3}
\title{Random Number Generation Function for 3-parameter twopiece distributions}
\usage{
rtp3(n, mu, par1, par2, FUN, param = "tp")
}
\arguments{
\item{x:}{vector of quantiles.}

\item{p:}{vector of probabilities.}

\item{n:}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{mu:}{location parameter.}

\item{par1:}{scale parameter 1.}

\item{par2:}{scale parameter 2.}

\item{FUN:}{a symmetric density f.}

\item{param:}{parameterizations used.}

\item{log, }{log.p: logical; if TRUE, probabilities p are given as log(p).}
}
\value{

}
\description{
Random Number Generation Function for 3-parameter twopiece distributions
}
